+++ truffle/src/com.oracle.truffle.api.instrumentation.test/src/com/oracle/truffle/api/instrumentation/test/ContextsEventsTest.java
@@ -265,12 +265,9 @@ public class ContextsEventsTest {
             try (Context context = Context.newBuilder().engine(engine).build()) {
                 context.eval(Source.create(InstrumentationTestLanguage.ID, "CONTEXT(STATEMENT())"));
                 for (ContextEvent event : events) {
-                    try {
-                        event.context.close();
-                        fail("Context close should fail.");
-                    } catch (UnsupportedOperationException ex) {
-                        // O.K.
-                    }
+                    // supported as long as not entered
+                    event.context.close();
+                    break;
                 }
             }
         }