+++ truffle/CHANGELOG.md
@@ -15,6 +15,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `TruffleContext.closeCancelled` and `TruffleContext.closeResourceExhausted`  to allow instruments and language that create inner contexts to cancel the execution of a context.
 * Added `TruffleContext.isActive` in addition to `TruffleContext.isEntered` and improved their documentation to indicate the difference.
 * Added `ContextsListener.onContextResetLimit` to allow instruments to listen to context limit reset events from the polyglot API.
+* All instances of `TruffleContext` accessible from instruments can now be closed by the instrument. Previously this was only possible for creators of the TruffleContext instance.
 
 ## Version 20.2.0
 * Added new internal engine option `ShowInternalStackFrames` to show internal frames specific to the language implementation in stack traces.